Refactor scripts and dockerfiles (#1)
* Make Scripts to specify Ubuntu Version

* Update BaseOS dockerfile

* Set ENV $USER

* Set Executable attribute

* Update Dockerfile-Ubuntu18.04

* Update Dockerfile-Ubuntu16.04

* Add wheel

* Update Dockerfile-OpenVINO-Toolkit

* Update Dockerfile-Ubuntu18.04

* Merge Base OS dockerfile

* Update Dockerfile-OpenVINO-Toolkit-Verification

* Update Dockerfile-Ubuntu

* Update Dockerfile-OpenVINO-Toolkit-Verification

* Update build-verify-container.sh

* Update Dockerfile-Ubuntu

* Update Dockerfile-Ubuntu

* Update build-BaseOS.sh

* Update build-BaseOS.sh

* Update Dockerfile-OpenVINO-Toolkit-Verification

* Update Dockerfile-OpenVINO-Toolkit-Verification

* Update Dockerfile-OpenVINO-Toolkit-Verification

* Update Dockerfile-OpenVINO-Toolkit-Verification

* Update Dockerfile-OpenVINO-Toolkit-Verification

* Update build-BaseOS.sh

* Add parameter check

* Update verify-container.sh

* Update verify-container.sh

* Update build-BaseOS.sh

* Update verify.sh

* update parameters

* update

* Add Benchmark App

* Update README for script changes

* Update Dockerfile-OpenVINO-Toolkit

* Update README.md

+++ Container/build-all-clean.sh
@@ -1,3 +1,14 @@
+if [ $# -ne 2 ]
+  then
+    echo "======================================="
+    echo "Please specify Ubuntu Version and reistry"
+    echo "  Ubuntu Version : 18.04 or 16.04"
+    echo "  Registry       : Your registry"
+    echo "  Example ./build-BaseOS.sh 18.04 myregistry"
+    echo "======================================="
+    exit
+fi
+
 docker system prune -a -f
-./build-BaseOS.sh
-./build-OpenVINO-Toolkit.sh
+./build-BaseOS.sh $1 $2
+./build-OpenVINO-Toolkit.sh $1 $2 "2019.3.376"
